+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Change Log
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## next
+* Fixed Vue `useFilter()` working with reactive filter (by Eduard Aksamitov).
+
 ## 0.14.4
 * Fixed logger for clean events (by Aleksandr Slepchenkov).
 * Updated `@nanostores/vue`.